requirements.txt
beautifulsoup4==4.10.0
chardet==3.0.4
gensim==3.8.3
google_patent_scraper==1.0.8
keras==2.7.0
lxml==4.5.2
matplotlib==3.2.2
nltk==3.5
numpy==1.19.4
pandas==1.3.3
Pillow==9.0.0
python_Levenshtein==0.12.2
PyYAML==6.0
scikit_criteria==0.2.10
scikit_learn==1.0.2
scipy==1.5.0
seaborn==0.11.2
simpletransformers==0.60.0
streamlit==1.2.0
textblob==0.17.1
XlsxWriter==1.2.9
tensorflow==2.7.0
torch==1.7.0